kernels: move Hub-related bits to `hf_hub` module
